Clac

Description

Documentation

data Opt

Holds information about the user passed options.

Constructors

MkOpt

Fields

wantHelp :: Bool

wantVerbose :: Bool

getEquation :: [String]

data StackItem a where

An item on the calculator stack. Can either be a constant number, or an operator.

Constructors

StackNum :: forall a. Fractional a => a -> StackItem a

StackOp :: OpDesc -> StackItem a

data OpDesc

Operator-description binding.

Constructors

MkOpDesc

Fields

op :: Op

desc :: String

data Op where

The valid operator classes.

Constructors

BinaryOp :: (forall a. Fractional a => a -> a -> a) -> Op

UnaryOp :: (forall a. Floating a => a -> a) -> Op

Constant :: (forall a. Floating a => a) -> Op

NewEq :: Op

optParser :: Parser Opt

Parser for the user requested options.

operators :: [(OpDesc, String)]

Predefined operator stack. This reflects the current program features.

buildStack

Arguments

:: String
one equation item

-> [StackItem Double]
prepend the result to this stack

-> [StackItem Double]

Build the calculator stack. Usually used in conjunction with foldr, to build the complete stack consecutively.

parseStack

Arguments

:: String
one equation item

-> Maybe (StackItem Double)
converted stack item, if valid

Parse an equation item.

stackTree :: [StackItem Double] -> Tree String

Build a tree from a calculator stack, suitable for pretty printing.

solveStack

Arguments

:: [StackItem Double]
stack to solve

-> [StackItem Double]
stack used for zipping, initially empty

-> Maybe Double
calculation result

Solve a stack.

solveAll :: [[String]] -> [(Maybe Double, String)]

Solve multiple equations of an input list such as:

 [["1","1","+"],["2","2","+"]]

This gives back a tuple for each equation where the first value is the result of the equation (if any) and the second one is a String representing the stackTree, suitable for printing.

solveEquation :: String -> Maybe Double

Solve an equation of the form 1 2 +.